FarmGen/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
appnope==0.1.4
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==24.2.0
blinker==1.9.0
boto3==1.35.56
botocore==1.35.56
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.7
decorator==5.1.1
Django==5.1.3
dnspython==2.7.0
exceptiongroup==1.2.2
executing==2.1.0
flatbuffers==24.3.25
fonttools==4.54.1
frozenlist==1.5.0
gast==0.6.0
gcloud==0.18.3
geographiclib==2.0
geopy==2.4.1
google-auth==2.36.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.67.1
h11==0.14.0
h5py==3.12.1
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
httpx-sse==0.4.0
idna==3.10
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.2
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jws==0.1.3
keras==2.15.0
kiwisolver==1.4.7
langchain==0.3.7
langchain-aws==0.2.7
langchain-community==0.3.5
langchain-core==0.3.15
langchain-mongodb==0.2.0
langchain-text-splitters==0.3.2
langsmith==0.1.142
libclang==18.1.1
Markdown==3.7
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
ml-dtypes==0.2.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.13.4
nest-asyncio==1.6.0
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
opencv-python==4.10.0.84
opt_einsum==3.4.0
orjson==3.10.11
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==4.25.5
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycryptodome==3.21.0
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
Pygments==2.18.0
pymongo==4.10.1
pyparsing==3.2.0
pypdf==5.1.0
Pyrebase4==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jwt==2.0.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==0.10.1
rpds-py==0.21.0
rsa==4.9
s3transfer==0.10.3
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
sqlparse==0.5.1
stack-data==0.6.3
tenacity==9.0.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
tf_keras==2.15.1
threadpoolctl==3.5.0
tornado==6.4.1
traitlets==5.14.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.14.1
yarl==1.17.1
